Refactor dashboard and templates pages; update sidebar and resume editor
- Updated sidebar and dashboard page components
- Removed old dashboard templates page and added new templates page
- Modified resume editor page
- Updated package-lock.json
